Complete rework of slider-popup system.
+ Completed backend for slider-popup system.
Started with forms, went to AJAX, and finally rested on plain old
JScript. The entire system has been finally redone so that it is
responsive and simple - Does not require any forms or AJAX commands. New
lightweight slider system implemented to allow for customization.

+ Popup Jscript modified, and will continue to be modified to allow for
anything to be popped up, in various styles of clicks.